+++ src/Products/PloneMeeting/profiles/__init__.py
@@ -570,6 +570,7 @@ class MeetingConfigDescriptor(Descriptor):
         self.adviceConfidentialFor = ()
         self.hideNotViewableLinkedItemsTo = ()
         self.inheritedAdviceRemoveableByAdviser = False
+        self.itemLabelsEditableByProposingGroupForever = False
         self.usingGroups = []
         # List of other meetingConfigs, item of this meetingConfig
         # will be clonable to